feat: migrate from Chakra UI to Tailwind CSS - complete migration while preserving look and feel
