+++ package.json
@@ -9,9 +9,6 @@
     "lint": "eslint"
   },
   "dependencies": {
-    "@chakra-ui/react": "^3.27.0",
-    "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.1",
     "@prisma/client": "^6.16.2",
     "@supabase/ssr": "^0.7.0",
     "@supabase/supabase-js": "^2.57.4",
