+++ src/components/controls/index.ts
@@ -1,3 +1,3 @@
 // Controls Components
-export { default as DeviceControlPanel } from './DeviceControlPanel';
-export { default as DeviceStatusCard } from './DeviceStatusCard';
+export { DeviceControlPanel } from './DeviceControlPanel';
+export { DeviceStatusCard } from './DeviceStatusCard';
