+++ src/hooks/useBreakpoint.ts
@@ -1,0 +1,30 @@
+'use client'
+
+import { useEffect, useState } from 'react'
+
+// Custom hook to replace Chakra's useBreakpointValue
+export function useBreakpoint() {
+  const [breakpoint, setBreakpoint] = useState<'sm' | 'md' | 'lg' | 'xl' | '2xl'>('lg')
+
+  useEffect(() => {
+    const handleResize = () => {
+      const width = window.innerWidth
+      if (width < 640) setBreakpoint('sm')
+      else if (width < 768) setBreakpoint('md')
+      else if (width < 1024) setBreakpoint('lg')
+      else if (width < 1280) setBreakpoint('xl')
+      else setBreakpoint('2xl')
+    }
+
+    handleResize()
+    window.addEventListener('resize', handleResize)
+    return () => window.removeEventListener('resize', handleResize)
+  }, [])
+
+  return breakpoint
+}
+
+export function useIsMobile() {
+  const breakpoint = useBreakpoint()
+  return breakpoint === 'sm' || breakpoint === 'md'
+}
